api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai==1.12.0

# Qdrant vector database
qdrant-client==1.7.3

# PostgreSQL
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and middleware
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
aiofiles==23.2.1
httpx==0.26.0
